project/src/components/offer-host/offer-host.tsx
import {Offer} from '../../types/offer';

type OfferHostProps = {
  offer: Offer;
};

function OfferHost({offer}: OfferHostProps): JSX.Element {
  const {host, description} = offer;

  return (
    <div className="property__host" data-testid="offer-host">
      <h2 className="property__host-title">Meet the host</h2>
      <div className="property__host-user user">
        <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
          <img
            className="property__avatar user__avatar"
            src={host.avatarUrl}
            width="74"
            height="74"
            alt={host.name}
          />
        </div>

        <span className="property__user-name">{host.name}</span>

        {host.isPro &&
        <span className="property__user-status">
          Pro
        </span>}
      </div>

      <div className="property__description">
        <p className="property__text">{description}</p>
      </div>
    </div>
  );
}

export default OfferHost;
